+++ contents-main.html
@@ -20,6 +20,7 @@
   <script src="js/service.intro.js"></script>
   <script src="js/service.license.js"></script>
   <script src="js/service.message.js"></script>
+  <script src="js/service.subtitle.js"></script>
   <script src="js/service.util.js"></script>
   <script src="js/services.main.js"></script>
   <script src="js/directive.include.js"></script>
