+++ contents-spec.html
@@ -16,6 +16,7 @@
   <script src="js/service.intro.js"></script>
   <script src="js/service.license.js"></script>
   <script src="js/service.message.js"></script>
+  <script src="js/service.subtitle.js"></script>
   <script src="js/service.util.js"></script>
   <script src="js/services.main.js"></script>
   <script src="js/service.aqusrdic.js"></script>
@@ -225,7 +226,7 @@
       <input id="record-result-2" type="text" ng-model="recordResult2">
     </div>
 
-    <div>AudioSourceService
+    <div>TextSubtitleService
       <input id="wav-file-path" type="text" ng-model="wavFilePath">
       <button id="source-fname" ng-click="ctrl.sourceFname()"></button>
       <input id="source-fname-result" type="text" ng-model="sourceFnameResult">
